HTML skeleton, setup w/ webpack
In this commit, I have undergone the setup process
for this repository, including a file structure
and configuration with webpack. I have also set up
an HTML skeleton to lay the foundation of my
project. The src js file includes a console.log
statement to ensure that "watch" and "build"
scripts are working as expected.